more dispose

+++ core/src/fi/tamk/tiko/PlayerProperties.java
@@ -1,8 +1,5 @@
 package fi.tamk.tiko;
 
-/**
- * Created by Omistaja on 6.3.2018.
- */
 
 public interface PlayerProperties {
     float normalSpeed = 2f;
@@ -11,6 +8,6 @@ public interface PlayerProperties {
     float playerHeight = 64;
     float moveSpeedXog = 200f;
     float moveSpeedYog = 200f;
-    float startX = 640f;
-    float startY = 300f;
+    int startX = 640;
+    int startY = 300;
 }
